pyproject.toml
[project]
name = "NerVault"
version = "1.0.0"
description = "A custodial web wallet for the Nerva (XNV) cryptocurrency."
license = "GPL-3.0"
authors = [
  { name="Sayan Bhattacharyya", email="sayan@sn1f3rt.dev" },
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "quart>=0.19.8",
    "quart-bcrypt>=0.0.9",
    "valkey>=6.0.2",
    "password-validator>=1.0",
    "aiohttp>=3.10.10",
    "docker>=7.1.0",
    "motor>=3.6.0",
    "quart-auth>=0.10.1",
    "six>=1.16.0",
    "quart-wtforms>=1.0.3",
    "qrcode[pil]>=8.0",
    "email-validator>=2.2.0",
    "aiosmtplib>=3.0.2",
    "httpx>=0.27.2",
    "nerva-py>=1.1.0",
    "python-dateutil>=2.9.0.post0",
]

[tool.uv]
dev-dependencies = [
    "pre-commit>=4.0.1",
    "ruff>=0.7.2",
]

[tool.ruff]
line-length = 85
target-version = "py313"

[tool.ruff.lint.isort]
section-order = ["future", "typing", "blueprints", "standard-library", "third-party", "first-party", "local-folder", "config"]
force-wrap-aliases = true
combine-as-imports = true
length-sort = true

[tool.ruff.lint.isort.sections]
"typing" = ["typing"]
"blueprints" = ["blueprints"]
"config" = ["config"]
